requirements.txt
appdirs==1.4.0
Django==1.10.5
django-crispy-forms==1.6.1
django-filter==1.0.1
django-markdown-deux==1.0.5
django-pagedown==0.1.1
djangorestframework==3.5.3
gunicorn==19.6.0
Markdown==2.6.7
markdown2==2.3.1
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.1.10
six==1.10.0
